refactor: type hints + utils consider TaxEntryEmpty

+++ src/QUI/ERP/Tax/TaxEntryEmpty.php
@@ -47,9 +47,9 @@ class TaxEntryEmpty extends QUI\QDOM
     /**
      * Return the vat value
      *
-     * @return integer
+     * @return integer|float
      */
-    public function getValue(): int
+    public function getValue(): float
     {
         return (int)0;
     }
